fix(routing): break router↔llm-cache circular import via decision.ts
Extract RoutingDecision to src/routing/decision.ts so llm-cache.ts can
import it without depending on router.ts. Remove unused nextTier variables
in tier-escalation.ts and dead makeStory helper in llm.test.ts.
