+++ src/routing/decision.ts
@@ -1,0 +1,9 @@
+import type { Complexity, ModelTier, TestStrategy } from "../config";
+
+/** Routing decision for a story */
+export interface RoutingDecision {
+  complexity: Complexity;
+  modelTier: ModelTier;
+  testStrategy: TestStrategy;
+  reasoning: string;
+}
